requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
psycopg2==2.9.9
python-dotenv==1.0.1